backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
openai==1.30.1
python-dotenv==1.0.1
sumy==0.11.0
nltk==3.8.1
pydantic==2.7.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
